"Add image upload for books and update related routes and controllers"
This diff adds the ability to upload an image for books in the application. The following changes were made:

1. The `multer` package was added to

+++ P7-Dev-Web-livres-main/backend/controllers/booksRoute.js
@@ -1,20 +1,70 @@
+
+const { unlink } = require("../app");
 const Book = require("../models/Books");
+const fs = require("fs");
 
 exports.createBook = (req, res, next) => {
-  console.log("les donner", req.body);
+  const BookObject = JSON.parse(req.body.book);
+  delete BookObject._id;
+  delete BookObject._userId;
   const book = new Book({
-    ...req.body,
+    ...BookObject,
+    userId: req.auth.userId,
+    imageUrl: `${req.protocol}://${req.get("host")}/images/${
+      req.file.filename
+    }`,
   });
+
   book
     .save()
-    .then(() => res.status(201).json({ message: "objet enregistré !" }))
-    .catch((error) => res.status(400).json({ error }));
+    .then(() => {
+      res.status(201).json({ message: "objet crée" });
+    })
+    .catch((error) => {
+      res.status(400).json({ error });
+    });
 };
 
 exports.modifyBook = (req, res, next) => {
-  Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.status(200).json({ message: "objet modifié" }))
-    .catch((error) => res.status(400).json({ error }));
+  Book.findById({ _id: req.params.id })
+    .then((book) => {
+      if (!book) {
+        return res.status(404).json({ message: "Livre introuvable" });
+      }
+
+      const userRating = book.ratings.find(
+        (rating) => rating.userId === req.auth.userId
+      );
+
+      if (!userRating) {
+        return res.status(401).json({ message: "Non autorisé" });
+      }
+
+      let bookObject;
+      if (req.file) {
+        bookObject = {
+          ...JSON.parse(req.body.book),
+          imageUrl: `${req.protocol}://${req.get("host")}/images/${
+            req.file.filename
+          }`,
+        };
+      } else {
+        bookObject = { ...req.body };
+        delete bookObject._userId;
+      }
+
+      Book.updateOne(
+        { _id: req.params.id },
+        { ...bookObject, _id: req.params.id }
+      )
+        .then(() => res.status(200).json({ message: "objet modifié" }))
+        .catch((error) => {
+          res.status(400).json({ error });
+        });
+    })
+    .catch((error) => {
+      res.status(400).json({ error });
+    });
 };
 
 exports.getAllBook = (req, res, next) => {
@@ -30,9 +80,34 @@ exports.getOneBook = (req, res, next) => {
 };
 
 exports.deleteBook = (req, res, next) => {
-  Book.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "objet supprimé" }))
-    .catch((error) => res.status(400).json({ error }));
+  Book.findById({ _id: req.params.id })
+    .then((book) => {
+      if (!book) {
+        return res.status(404).json({ message: "Livre introuvable" });
+      }
+
+      const userRating = book.ratings.find(
+        (rating) => rating.userId === req.auth.userId
+      );
+
+      if (!userRating) {
+        return res.status(401).json({ message: "Non autorisé" });
+      }
+
+      const filename = book.imageUrl.split("/images/")[1];
+      fs.unlink(`images/${filename}`, () => {
+        Book.deleteOne({ _id: req.params.id })
+          .then(() => {
+            res.status(200).json({ message: "objet supprimé !" });
+          })
+          .catch((error) => {
+            res.status(500).json({ error });
+          });
+      });
+    })
+    .catch((error) => {
+      res.status(500).json({ error });
+    });
 };
 
 exports.getBestRatedBooks = (req, res, next) => {
@@ -47,19 +122,20 @@ exports.getBestRatedBooks = (req, res, next) => {
     });
 };
 
-// controllers/booksRoute.js
+
 exports.rateBook = (req, res, next) => {
   const userId = req.body.userId;
   const rating = req.body.rating;
+  const bookId = req.params.id;
 
-  // Assurez-vous que la note est entre 0 et 5
+
   if (rating < 0 || rating > 5) {
     return res
       .status(400)
       .json({ error: "La note doit être comprise entre 0 et 5." });
   }
 
-  // Recherchez le livre par son ID
+
   Book.findById(req.params.id)
     .then((book) => {
       if (!book) {
@@ -80,7 +156,7 @@ exports.rateBook = (req, res, next) => {
       book.rating.push(newRating);
       book.averageRating = calculateAverageRating(book.rating);
 
-      // Sauvegardez les modifications
+
       return book.save();
     })
     .then((updatedBook) => {
